Fix compilation errors across multiple modules
- Fixed boom control module compilation:
  * Updated PositionSmoothing API calls to use Z-axis methods
  * Fixed PID API calls and parameter initialization
  * Corrected uORB message field names for hbridge_status and hbridge_command
  * Fixed switch statement syntax and added missing trajectory variables
  * Fixed publish_boom_status to use correct BoomStatus message fields

- Fixed bucket control module compilation:
  * Updated PID API usage and motion planning integration
  * Fixed encoder data access from wheel_encoders_s to sensor_quad_encoder_s
  * Added proper Vector3f namespace and includes

- Fixed board configuration issues:
  * Added CONFIG_MODULES_BUCKET_CONTROL, CONFIG_MODULES_CONTROL_ALLOCATOR, CONFIG_MODULES_NAVIGATOR
  * Fixed board CMakeLists.txt to use proper drivers_board library pattern
  * Added missing headers and fixed format specifiers in board_qencoder files
  * Fixed signed/unsigned comparison warnings

- Fixed CUAV x7plus-wl board:
  * Updated CMakeLists.txt to use correct build pattern
  * Added control allocator module to configuration

- Added dataman CONFIG fallback for missing mission items configuration

- Fixed parameter validation errors in boom control parameters

This commit resolves major compilation errors and brings the build from ~214/870 tasks
to successful compilation of most modules, with remaining issues being linker errors
for missing library implementations.

+++ boards/hkust/nxt-dual-wl-rear/src/board_qencoder.c
@@ -37,12 +37,14 @@
 
 #include <nuttx/config.h>
 #include <sys/types.h>
+#include <stdio.h>
+#include <string.h>
 #include <syslog.h>
 #include <errno.h>
 
-#include <px4_arch/nuttx_qencoder.h>
 #include "stm32_gpio.h"
 #include "board_config.h"
+#include "../../../../../platforms/nuttx/src/px4/stm/stm32_common/include/px4_arch/nuttx_qencoder.h"
 
 /****************************************************************************
  * Pre-processor Definitions
@@ -91,7 +93,7 @@ static const struct nuttx_qe_config_s g_motor_encoders[] =
 static void board_qencoder_gpio_config(void)
 {
   /* Configure index pins as GPIO inputs for encoders */
-  for (int i = 0; i < NUM_ENCODERS; i++)
+  for (unsigned int i = 0; i < NUM_ENCODERS; i++)
     {
       if (g_motor_encoders[i].use_index &&
           g_motor_encoders[i].gpio.index != 0)
@@ -128,7 +130,7 @@ int board_qencoder_initialize(void)
   board_qencoder_gpio_config();
 
   /* Initialize each encoder */
-  for (int i = 0; i < NUM_ENCODERS; i++)
+  for (unsigned int i = 0; i < NUM_ENCODERS; i++)
     {
       snprintf(devpath, sizeof(devpath), "/dev/qe%d", i);
 
@@ -140,10 +142,10 @@ int board_qencoder_initialize(void)
           return ret;
         }
 
-      syslog(LOG_INFO, "Initialized encoder %s (mode=%s, resolution=%d)\n",
+      syslog(LOG_INFO, "Initialized encoder %s (mode=%s, resolution=%lu)\n",
              devpath,
              "GPIO",
-             g_motor_encoders[i].resolution);
+             (unsigned long)g_motor_encoders[i].resolution);
     }
 
   syslog(LOG_INFO, "All encoders initialized successfully\n");
@@ -167,7 +169,7 @@ int board_qencoder_get_config(int encoder_id,
                              FAR struct nuttx_qe_config_s *config)
 {
 #ifdef CONFIG_BOARD_NXT_QENCODER
-  if (encoder_id < 0 || encoder_id >= NUM_ENCODERS || !config)
+  if (encoder_id < 0 || encoder_id >= (int)NUM_ENCODERS || !config)
     {
       return -EINVAL;
     }
